Apply the approval-round nits: honest copy, truncation surfaced, a11y states
