+++ src/translations/en.json
@@ -1287,12 +1287,14 @@
     "device_includes_heading": "Picking a device pre-fills the GitHub report with:",
     "device_includes_config": "Its YAML config, known credentials masked",
     "device_includes_facts": "Device details: name, board, platform, versions, installation",
-    "device_hint": "Nothing is sent until you review and submit the form.",
+    "device_hint": "Nothing is filed until you review and submit the form.",
     "device_skip": "No specific device",
     "device_filter": "Filter devices",
     "device_no_matches": "No devices match your filter",
+    "device_none": "No devices are configured yet",
     "device_capture_failed": "Couldn't capture the device config; the form opens without it",
     "device_ready": "A pre-filled GitHub issue opened in a new tab. If it didn't, use the button below.",
+    "device_ready_truncated": "A pre-filled GitHub issue opened in a new tab; the config was trimmed to fit, so paste the rest into the form if needed. If the tab didn't open, use the button below.",
     "open_form": "Open GitHub issue",
     "back": "Back",
     "browse_builder": "Device Builder issues",
